Adding game_utils_test to make it a little easier to find things.
